+++ Source/Android/Harvest/app/src/main/java/za/org/samac/harvest/SignIn_Choose.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 import android.os.PersistableBundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.view.View;
 
 public class SignIn_Choose extends AppCompatActivity {
@@ -31,5 +32,15 @@ public class SignIn_Choose extends AppCompatActivity {
         Intent openRegistration = new Intent(SignIn_Choose.this, SignIn_SignUp.class);
         openRegistration.setFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
         startActivityIfNeeded(openRegistration, 0);
+    }
+
+    //Back acts as home
+    @Override
+    public void onBackPressed() {
+        Log.d("CDA", "onBackPressed Called");
+        Intent setIntent = new Intent(Intent.ACTION_MAIN);
+        setIntent.addCategory(Intent.CATEGORY_HOME);
+        setIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        startActivity(setIntent);
     }
 }
