+++ Source/Android/Harvest/app/src/main/java/za/org/samac/harvest/SignIn_Farmer.java
@@ -243,15 +243,6 @@ public class SignIn_Farmer extends AppCompatActivity implements  GoogleApiClient
         //updateUI(account);
     }
 
-    @Override
-    public void onBackPressed() {
-        Log.d("CDA", "onBackPressed Called");
-        Intent setIntent = new Intent(Intent.ACTION_MAIN);
-        setIntent.addCategory(Intent.CATEGORY_HOME);
-        setIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        startActivity(setIntent);
-    }
-
     private void signInWithGoogle() {
         Intent signInIntent = mGoogleSignInClient.getSignInIntent();
         startActivityForResult(signInIntent, RC_SIGN_IN);
